+++ .changeset/move-npm-trust-commands.md
@@ -1,0 +1,8 @@
+---
+monochange: patch
+monochange_npm: minor
+---
+
+# Move npm publish helpers into monochange_npm
+
+Move npm trusted-publishing command helpers and npm placeholder manifest generation out of `monochange::package_publish` and into `monochange_npm`.
