+++ crates/monochange_npm/Cargo.toml
@@ -16,7 +16,9 @@ description = "Manage versions and releases for your multiplatform, multilanguag
 glob = { workspace = true, default-features = true }
 monochange_core = { workspace = true }
 monochange_ecmascript = { workspace = true }
+monochange_github = { workspace = true }
 monochange_linting = { workspace = true }
+monochange_publish = { workspace = true }
 semver = { workspace = true, default-features = true }
 serde = { workspace = true, default-features = true }
 serde_json = { workspace = true, default-features = true }
